+++ src/admin/testimonial/Testimonial.jsx
@@ -45,7 +45,7 @@ const Testimonial = ({ DataContext, FunctionContext }) => {
 
     const handleChangeStatus = async (id) => {
         try {
-
+            setIsFormProcess({ status: true, _id: id });
             let raw = await fetch(`${backendHost}/testimonial/update-status/${id}`, {
                 method: 'GET',
                 headers: {
@@ -98,48 +98,54 @@ const Testimonial = ({ DataContext, FunctionContext }) => {
     }
 
     const handleDeleteTestimonial = async (id) => {
-        try {
 
-            let raw = await fetch(`${backendHost}/testimonial/${id}`, {
-                method: 'DELETE',
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            })
+        let confirmation = window.confirm('Are you sure want to delete page')
 
-            let data = await raw.json()
+        if (confirmation === true) {
+            setIsFormProcess({ status: true, _id: id });
+            try {
 
-            setTestimonialData(testimonialData.filter((ele) => ele._id !== id))
+                let raw = await fetch(`${backendHost}/testimonial/${id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    }
+                })
 
-            setResponseStatus(true)
-            setResponseData({
-                isLoading: false,
-                heading: "Information",
-                message: data
-            })
+                let data = await raw.json()
 
-            setIsFormProcess({ _id: "", status: false })
+                setTestimonialData(testimonialData.filter((ele) => ele._id !== id))
+
+                setResponseStatus(true)
+                setResponseData({
+                    isLoading: false,
+                    heading: "Information",
+                    message: data
+                })
+
+                setIsFormProcess({ _id: "", status: false })
+                setTimeout(() => {
+                    setResponseStatus(false)
+                }, 2000);
+
+            } catch (error) {
+                setResponseStatus(true)
+                setResponseData({
+                    isLoading: false,
+                    heading: `Server Error`,
+                    message: error.message
+                })
+            }
+
             setTimeout(() => {
                 setResponseStatus(false)
-            }, 2000);
-
-        } catch (error) {
-            setResponseStatus(true)
-            setResponseData({
-                isLoading: false,
-                heading: `Server Error`,
-                message: error.message
-            })
+                setResponseData({
+                    isLoading: false,
+                    heading: "",
+                    message: ""
+                })
+            }, 4000);
         }
-
-        setTimeout(() => {
-            setResponseStatus(false)
-            setResponseData({
-                isLoading: false,
-                heading: "",
-                message: ""
-            })
-        }, 4000);
     }
 
     return (
@@ -152,8 +158,8 @@ const Testimonial = ({ DataContext, FunctionContext }) => {
                     <span>Loading Data</span>
                 </div>
                 :
-                <div className="w-100">
-                    <table className='w-100 mt-5 border border-theam mb-auto'>
+                <div className="w-100 py-md-5 py-3">
+                    <table className='w-100 border border-theam mb-auto'>
                         <thead>
                             <tr className='bg-theam'>
                                 <th className='py-2 text-white border-end text-center px-2'>Date</th>
@@ -181,7 +187,7 @@ const Testimonial = ({ DataContext, FunctionContext }) => {
                                                     </>
                                                     :
                                                     <>
-                                                        <button type="button" className='btn-reset lh-1 bg-theam p-2 rounded-circle' onClick={() => { setIsFormProcess({ status: true, _id: ele._id }); handleChangeStatus(ele._id) }}>
+                                                        <button type="button" className='btn-reset lh-1 bg-theam p-2 rounded-circle' onClick={() => { handleChangeStatus(ele._id) }}>
                                                             {ele.status === true ?
                                                                 <FaEye className='text-white' />
                                                                 :
@@ -189,7 +195,7 @@ const Testimonial = ({ DataContext, FunctionContext }) => {
                                                             }
                                                         </button>
 
-                                                        <button type="button" className='btn-reset lh-1 bg-danger p-2 rounded-circle' onClick={() => { setIsFormProcess({ status: true, _id: ele._id }); handleDeleteTestimonial(ele._id) }}>
+                                                        <button type="button" className='btn-reset lh-1 bg-danger p-2 rounded-circle' onClick={() => { handleDeleteTestimonial(ele._id) }}>
                                                             <FaTrash className='text-white' />
                                                         </button>
                                                     </>
